+++ org.eclipse.winery.repository.ui/src/app/instance/serviceTemplates/boundaryDefinitions/propertyMappings/propertyMappings.component.ts
@@ -30,9 +30,9 @@ export class PropertyMappingsComponent implements OnInit {
     loading = true;
     apiData: PropertyMappingsApiData;
     columns: Array<WineryTableColumn> = [
-        {title: 'Service Template Property', name: 'serviceTemplatePropertyRef', sort: true},
-        {title: 'Target', name: 'targetObjectRef', sort: true},
-        {title: 'Target Property', name: 'targetPropertyRef', sort: true}
+        { title: 'Service Template Property', name: 'serviceTemplatePropertyRef', sort: true },
+        { title: 'Target', name: 'targetObjectRef', sort: true },
+        { title: 'Target Property', name: 'targetPropertyRef', sort: true }
     ];
     @ViewChild('addPropertyMappingModal') addPropertyMappingModal: ModalDirective;
     @ViewChild('confirmDeleteModal') confirmDeleteModal: ModalDirective;
@@ -42,7 +42,7 @@ export class PropertyMappingsComponent implements OnInit {
     @ViewChild('propertiesSelect') propertiesSelect: any;
     currentSelectedItem: Property = new Property();
     addOrUpdateBtnTxt = 'Add';
-    properties: { name: string, property: string } = {name: '', property: ''};
+    properties: { name: string, property: string } = { name: '', property: '' };
     xmlData: any;
     selectedProperty: string;
     templateList: Array<any> = [];
@@ -51,7 +51,7 @@ export class PropertyMappingsComponent implements OnInit {
     targetProperties: Array<any> = [];
     targetObject: string;
     targetPropertiesWrapperElement: string = null;
-    initialSelectProp: any = [{id: '', text: ''}];
+    initialSelectProp: any = [{ id: '', text: '' }];
     initialSelectedTempalte: [{ id: '', text: '' }];
 
     constructor(private service: PropertyMappingService,
@@ -101,8 +101,8 @@ export class PropertyMappingsComponent implements OnInit {
         if (!isNullOrUndefined(this.topologyTemplate[templateType])) {
             return this.topologyTemplate[templateType].map((template: WineryTemplate) => {
                 const newItem: SelectItem = new SelectItem('');
-                newItem.id = template.type;
-                newItem.text = Utils.getNameFromQname(template.type);
+                newItem.id = template.id;
+                newItem.text = Utils.getNameFromQname(template.id);
                 return newItem;
             });
         } else {
@@ -113,7 +113,6 @@ export class PropertyMappingsComponent implements OnInit {
     }
 
     targetObjectSelected(targetObj: any) {
-        console.log(targetObj);
         this.targetObject = targetObj.text;
         this.currentSelectedItem.targetObjectRef = this.targetObject;
         this.getTargetProperties(targetObj);
@@ -137,7 +136,7 @@ export class PropertyMappingsComponent implements OnInit {
     handleGetProperties(propertiesDefinition: PropertiesDefinitionsResourceApiData) {
         if (!isNullOrUndefined(propertiesDefinition.winerysPropertiesDefinition)) {
             this.targetProperties = propertiesDefinition.winerysPropertiesDefinition.propertyDefinitionKVList.map(item => {
-                return {id: item.key, text: item.key}
+                return { id: item.key, text: item.key }
             });
             this.targetPropertiesWrapperElement = propertiesDefinition.winerysPropertiesDefinition.elementName;
             this.initialSelectProp = [this.targetProperties[0]];