+++ CHANGELOG.md
@@ -15,7 +15,8 @@ This is similar to [Angular's CHANGELOG.md](https://github.com/angular/angular/b
 ## [unreleased]
 
 ### Changed
-- Add ZIP-button for artifacttemplate-files
+- Add available/not available indicator for LFS in git log component
+- Add ZIP-button for artifactemplate-files
 - Add license and readme support for all components
 - Fix delete dialog message text
 - Fix popup text of upload message